app/requirements.txt
fastapi
uvicorn
transformers
torch
torchvision
Pillow
python-multipart
python-dotenv
timm
